Implement user authorization logic
Add cookie-parser to dependencies.

User token created once the user gets authenticated and logs in.
This is a jsonwebtoken that gets stored as a cookie.

Create a middleware to authorize a user the access. This consist in
retrieving the token cookie previously created in the login route and
verify it's validity. If the token is valid, the user is authorized.

Create a route to send the user information once it's authorized.