Initial face-detection post and put handlers
Install multer package as a dependency and import default modules fs,
crypto in order to process multipart/form-data (form with files) and be
able to recieve an image from the front end and store it with a unique
name. This file will then be sent to the Clarifai celebrity recognition
API and the response will then be handled.

The post handler recieves and processess the image. The put handler will
increment the user entry count in the database.