Implement connection and auth for Clarifai API
Connection now allows to send an image URL to Clarifai API and recieve
a response containing the data from it's celebrity-face-detection IA
model.

The important data is sent to the front-end (boundingBox coordinates,
name and correct detection probability).

IMPORTANT: Local development doesn't allow to serve images from the
local server, so the image URL for the request is hard coded. Must
change to a dinamic URL (already done but commented out) once the
server is online.
All uploading images and serving them logic is done.